requirements.txt
accelerate==0.33.0
bitsandbytes==0.43.2
ExifRead==3.0.0
imagesize==1.4.1
pillow==10.4.0
pyparsing==3.1.2
PySide6==6.7.2
# Transformers v4.42 breaks CogVLM.
transformers==4.41.2

# PyTorch
torch==2.2.2; platform_system != "Windows"
https://download.pytorch.org/whl/cu121/torch-2.2.2%2Bcu121-cp311-cp311-win_amd64.whl; platform_system == "Windows" and python_version == "3.11"
https://download.pytorch.org/whl/cu121/torch-2.2.2%2Bcu121-cp310-cp310-win_amd64.whl; platform_system == "Windows" and python_version == "3.10"

# FlashAttention (Florence-2, Phi-3-Vision)
flash-attn==2.6.3; platform_system == "Linux"
https://github.com/oobabooga/flash-attention/releases/download/v2.6.1/flash_attn-2.6.1+cu122torch2.2.2cxx11abiFALSE-cp311-cp311-win_amd64.whl; platform_system == "Windows" and python_version == "3.11"
https://github.com/oobabooga/flash-attention/releases/download/v2.6.1/flash_attn-2.6.1+cu122torch2.2.2cxx11abiFALSE-cp310-cp310-win_amd64.whl; platform_system == "Windows" and python_version == "3.10"

# CogAgent
timm==1.0.7

nvidia-ml-py==12.535.77  # For GPU monitoring



python-dotenv==1.0.0

# CogVLM
einops==0.8.0
protobuf==5.27.2
sentencepiece==0.2.0
torchvision==0.17.2
xformers==0.0.25.post1

# InternLM-XComposer2
auto-gptq==0.7.1; platform_system == "Linux" or platform_system == "Windows"
numpy==1.26.4

# WD Tagger
huggingface-hub==0.24.2
onnxruntime==1.18.1
